+++ src/pinocchio.h
@@ -48,7 +48,7 @@ string int_left(string s, unsigned int w, char c);
 string int_right(string s, unsigned int w, char c);
 string int_a2s(void *x);
 string int_strip(string s, char side, string w);
-size_t int_cindex(string s, char c, signed long long int n, size_t start);
+size_t int_cindex(string s, char c, int n, size_t start);
 size_t int_ccount(string s, char c, size_t start, size_t end);
 string int_creplace(string s, char c, char r, size_t n, size_t start, size_t end);
 size_t int_scount(string s, string ss/*, size_t start, size_t end*/);
